When events are found, async requests are made to save them in our database (our format, not api's format). Key in DB is apiName + id of event from API.
Changed longitude to be so that it is passed as a signed number in the api request (-72.9 for new haven) as our api worked for eventbrite but not ticketmaster if passed a negative number or vice versa iif passed a positive number.

Some API requests from eventbrite/ticketmaster didn't have starttime, so I defaulted them to null. Have to see if this is a mandatory item for us though, if so we will remove such events. (Will talk in standup)